backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0